moved mockData  (#207)
* moved mockData to share

* Add generic path to mockdata

* change references to generic references

---------
