+++ cms/src/components/content/EditContentVideo.spec.ts
@@ -3,7 +3,7 @@ import { describe, it, expect, vi, beforeAll, afterAll } from "vitest";
 import { mount } from "@vue/test-utils";
 import { createTestingPinia } from "@pinia/testing";
 import { type ContentDto, accessMap } from "luminary-shared";
-import * as mockData from "@/tests/mockData";
+import * as mockData from "@mockdata";
 import { setActivePinia } from "pinia";
 import { ref } from "vue";
 import EditContentVideo from "./EditContentVideo.vue";
