+++ cms/tsconfig.app.json
@@ -1,13 +1,14 @@
 {
     "extends": "@vue/tsconfig/tsconfig.dom.json",
-    "include": ["env.d.ts", "src/**/*", "src/**/*.vue"],
+    "include": ["env.d.ts", "src/**/*", "src/**/*.vue", "mockdata/**/*"],
     "exclude": ["src/**/*.spec.ts"],
     "compilerOptions": {
         "composite": true,
         "noEmit": true,
         "baseUrl": ".",
         "paths": {
-            "@/*": ["./src/*"]
+            "@/*": ["./src/*"],
+            "@mockdata": ["../mockdata"]
         }
     }
 }
