+++ mockdata/README.md
@@ -1,0 +1,3 @@
+# Mock Data
+
+This directory contains mock data used for unit tests for the CMS and app in this mono-repo.
